попытка сделать фильтр
